JAV-469 rename InstanceCacheManager to InstanceCacheManagerOld, and create interface InstanceCacheManager
